Python/requirements.txt
asgiref==3.2.7
click==7.1.2
Django==3.0.6
django-filter==2.2.0
djangorestframework==3.11.0
dlib==19.19.0
face-recognition==1.3.0
face-recognition-models==0.3.0
importlib-metadata==1.6.0
imutils==0.5.3
Markdown==3.2.2
numpy==1.18.4
opencv-python==4.2.0.34
pickle-mixin==1.0.2
Pillow==7.1.2
psycopg2-binary==2.8.5
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1
zipp==3.1.0
